requirements.txt
# requirements.txt

fastapi
uvicorn
typer[all]
openapi-spec-validator
pydantic
httpx
sqlalchemy
jinja2
pyyaml
pytest
psycopg2-binary
pytest-asyncio
numpy
matplotlib
scikit-learn
sdv
